Update README.md
Reordered the Readme to more accurately represent the timeline of our analysis. Added a Summary/Abstract at the start, which needs to be completed with final analysis. Perhaps a brief description of our models.
Analysis section at the end also needs to be completed, also filled with details of our analysis.
Conclusion section still needs to be added.